requirements.txt
aiofiles
fastapi
uvicorn
python-dotenv
pydantic
# --- AI / ML (LangChain Ecosystem) ---
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-chroma
langchain-openai
langchain-text-splitters
chromadb
sentence-transformers
openai
torch
transformers
# --- Utilities ---
tqdm
pandas
beautifulsoup4
python-frontmatter
lxml
tabulate
psutil
einops